# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'API Documentation'
  description: "Documentation des API de l'application Les Innovateurs. Cette documentation fournit des informations détaillées sur les différentes API disponibles pour gérer les utilisateurs, les élections, les candidats, les activités, les sondages et d'autres fonctionnalités de l'application."
  license:
    name: 'URL Accueil'
    url: 'https://lesinnovateurs.me'
  version: 1.0.0
paths:
...
  not-defined-yet:
    get:
      tags:
        - 'Candidat Authenticated Actions'
      summary: "Récupérer  les  programmes d'un candidat"
      description: "Récupérer les  programmes d'un candidat"
      operationId: 7461c3a2db6a133a04477e68d63b5c62
      parameters:
        -
          name: id
          in: path
          description: 'ID du candidat lié au programme'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: 'Programme récupéré avec succès'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: string
                    example: 'Programme récupéré'
                  data:
                    $ref: '#/components/schemas/Programme'
                type: object
        404:
          description: 'Programme non trouvé'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Programme non trouvé'
                type: object
    put:
      tags:
        - 'User Authenticated actions'
      summary: 'Modifier un commentaire'
      operationId: 4659c27c3500950dfec64a6e0529615a
      parameters:
        -
          name: id
          in: path
          description: 'Id du commentaire'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Commentaire'
      responses:
        200:
          description: 'Commentaire modifie avec succes'
        404:
          description: 'Commentaire non trouve'
    post:
      tags:
        - 'Admin Actions'
      summary: 'Add a new admin'
      operationId: 47e8ba5e07df7c220e55ec1dd2614b7d
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Admin'
      responses:
        201:
          description: 'Admin added successfully'
        400:
          description: 'Bad request'
    delete:
      tags:
        - 'User Authenticated actions'
      summary: 'Supprimer un vote'
      operationId: 401492d0ee002b21130078492f0a9435
      parameters:
        -
          name: id_election
          in: path
          description: "Id de l'election"
          required: true
          schema:
            type: integer
            format: int64
        -
          name: id_user
          in: path
          description: "Id de l'utilisateur"
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: 'Vote supprime avec succes'
        404:
          description: 'Vote non trouve'
...
  /not-defined-yet:
    post:
      tags:
        - 'Admin Actions'
      summary: 'Ajouter un candidat'
      operationId: e23b3611a2b9771d3c63e4de23ed5472
      parameters:
        -
          name: id_user
          in: path
          description: "ID de l'utilisateur à enregistrer comme candidat"
          required: true
          schema:
            type: integer
        -
          name: id_parti_politique
          in: path
          description: 'ID du parti politique auquel le candidat appartient'
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Candidat ajouté avec succès'
        404:
          description: 'Utilisateur non trouvé'
...
  /api/private/candidat/add-programme:
    post:
      tags:
        - 'Candidat Authenticated Actions'
      summary: 'Ajouter un programme'
      description: 'Ajouter un programme'
      operationId: 04c6e44508caa6c83113a3f143285b74
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Programme'
      responses:
        201:
          description: 'Programme ajouté avec succès'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: string
                    example: 'Programme ajouté'
                  programme:
                    $ref: '#/components/schemas/Programme'
                type: object
        400:
          description: 'Bad request'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Vous avez deja ajouté ce programme'
                type: object
  '/api/public/programme/{val}':
    get:
      tags:
        - 'Candidat Authenticated Actions'
      summary: 'Rechercher un programme'
      description: 'Rechercher un programme'
      operationId: b52d58eb13df43723440c455088c0982
      parameters:
        -
          name: val
          in: path
          description: 'valeur de la recherche du programme'
          required: true
          schema:
            type: string
            format: string
      responses:
        200:
          description: 'Programme récupéré avec succès'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: string
                    example: 'Programme récupéré'
                  data:
                    $ref: '#/components/schemas/Programme'
                type: object
        404:
          description: 'Programme non trouvé'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Programme non trouvé'
                type: object
  '/api/private/candidat/update-programme/{id}':
    put:
      tags:
        - 'Candidat Authenticated Actions'
      summary: 'Modifier un programme'
      description: 'Modifier un programme'
      operationId: 71d9ba4a8f3da64b6595871722775d3f
      parameters:
        -
          name: id
          in: path
          description: 'ID du programme'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Programme'
      responses:
        200:
          description: 'Programme modifié avec succès'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: string
                    example: 'Programme modifié'
                type: object
        404:
          description: 'Programme non trouvé'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Programme non trouvé'
                type: object
  '/api/private/candidat/delete-programme/{id}':
    delete:
      tags:
        - 'Candidat Authenticated Actions'
      summary: 'Supprimer un programme'
      description: 'Supprimer un programme'
      operationId: 24bcdae77f0805eb3d1d8da7acd512d1
      parameters:
        -
          name: id
          in: path
          description: 'ID du programme'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: 'Programme supprimé avec succès'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: string
                    example: 'Programme supprimé'
                type: object
        404:
          description: 'Programme non trouvé'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Programme non trouvé'
                type: object
  /api/private/candidat/my-programmes:
    get:
      tags:
        - 'Candidat Authenticated Actions'
      summary: "Récupérer tous les programmes d'un candidat"
      description: "Récupérer tous les programmes d'un candidat"
      operationId: d7ed19a77e50729a27d898db21431c8d
      responses:
        200:
          description: 'Programme récupéré avec succès'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: string
                    example: 'Programme récupéré'
                  data:
                    $ref: '#/components/schemas/Programme'
                type: object
        404:
          description: 'Programme non trouvé'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Programme non trouvé'
                type: object
  /api/private/user/programmes-not-reacted:
    get:
      tags:
        - 'User Authenticated Actions'
      summary: "Récupérer tous les programmes d'un candidat"
      description: "Récupérer tous les programmes d'un candidat"
      operationId: be9e33143507814af3252180f3804f1b
      parameters:
        -
          name: id_candidat
          in: path
          description: 'ID du candidat lié au programme'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: 'Programme récupéré avec succès'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: string
                    example: 'Programme récupéré'
                  data:
                    $ref: '#/components/schemas/Programme'
                type: object
        404:
          description: 'Programme non trouvé'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Programme non trouvé'
                type: object
  '/api/public/all-programmes/{id_candidat}':
    get:
      tags:
        - 'Public API'
      summary: "Récupérer tous les programmes d'un candidat"
      description: "Récupérer tous les programmes d'un candidat"
      operationId: 6b9451e4a40bf27b2ccb4ba42b2b326c
      parameters:
        -
          name: id_candidat
          in: path
          description: 'ID du candidat lié au programme'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: 'Programme récupéré avec succès'
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: string
                    example: 'Programme récupéré'
                  data:
                    $ref: '#/components/schemas/Programme'
                type: object
        404:
          description: 'Programme non trouvé'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Programme non trouvé'
                type: object
...
components:
  schemas:
...
    Admin:
      title: Admin
      description: 'Admin model'
      properties:
        id_user:
          description: 'User ID associated with the admin'
          type: integer
        created_at:
          description: 'Admin creation date and time'
          type: string
          format: date-time
          readOnly: 'true'
        updated_at:
          description: 'Admin last update date and time'
          type: string
          format: date-time
          readOnly: 'true'
      type: object
...
    Commentaire:
      title: Commentaire
      description: 'Commentaire model'
      properties:
        id_user:
          description: 'User ID associated with the comment'
          type: integer
        id_post:
          description: 'ID of the post associated with the comment'
          type: integer
        commentaire:
          description: 'Content of the comment'
          type: string
        created_at:
          description: 'Comment creation date and time'
          type: string
          format: date-time
          readOnly: 'true'
        updated_at:
          description: 'Comment last update date and time'
          type: string
          format: date-time
          readOnly: 'true'
      type: object
...
    Programme:
      title: Programme
      description: 'Modèle de programme'
      properties:
        id_candidat:
          description: 'ID du candidat lié au programme'
          type: integer
        titre:
          description: 'Titre du programme'
          type: string
        description:
          description: 'Description du programme'
          type: string
        url_media:
          description: 'URL du média associé au programme'
          type: string
        created_at:
          description: 'Date et heure de création'
          type: string
          format: date-time
          readOnly: 'true'
          nullable: true
        updated_at:
          description: 'Date et heure de mise à jour'
          type: string
          format: date-time
          readOnly: 'true'
          nullable: true
      type: object
...